Let a tracked product choose its unit and hold decimal thresholds
Stock can be fractional, but the product form had no way to say what one unit is, and it read
reorder points and overstock thresholds as whole numbers, so 2.5 saved as zero. Tracked products
now have a unit dropdown, and both thresholds are read as decimals.

+++ ArgoBooks/ViewModels/ProductsPageViewModel.cs
@@ -726,8 +726,8 @@ public partial class ProductsPageViewModel : SortablePageViewModelBase
         companyData.IdCounters.Product++;
         var newId = $"PRD-{companyData.IdCounters.Product:D3}";
 
-        var reorderPoint = int.TryParse(ModalReorderPoint, out var rp) ? rp : 0;
-        var overstockThreshold = int.TryParse(ModalOverstockThreshold, out var ot) ? ot : 0;
+        var reorderPoint = decimal.TryParse(ModalReorderPoint, out var rp) ? rp : 0;
+        var overstockThreshold = decimal.TryParse(ModalOverstockThreshold, out var ot) ? ot : 0;
 
         var newProduct = new Product
         {
@@ -830,8 +830,8 @@ public partial class ProductsPageViewModel : SortablePageViewModelBase
         var newSupplierId = ModalSupplier?.Id;
         var newUnitPrice = decimal.TryParse(ModalUnitPrice, out var unitPrice) ? unitPrice : 0;
         var newCostPrice = decimal.TryParse(ModalCostPrice, out var costPrice) ? costPrice : 0;
-        var newReorderPoint = int.TryParse(ModalReorderPoint, out var rp) ? rp : 0;
-        var newOverstockThreshold = int.TryParse(ModalOverstockThreshold, out var ot) ? ot : 0;
+        var newReorderPoint = decimal.TryParse(ModalReorderPoint, out var rp) ? rp : 0;
+        var newOverstockThreshold = decimal.TryParse(ModalOverstockThreshold, out var ot) ? ot : 0;
         var newTrackInventory = ModalTrackInventory;
 
         // Update the product
